feat: concentrate eigenlayer avs interactions into service manager (#109)
* feat: reintroduce IServiceManager & ServiceManagerBase

These are useful concepts to form the single interaction point
where AVSs push data to EigenLayer.
The previous version focused on (unused) Slashing interactions,
whereas this version focused on the newer registration interactions.
TODO: modify the RegistryCoordinator to call the ServiceManagerBase
instead of calling the DelegationManager directly.

* feat: integrate RegistryCoordinator with ServiceManager(Base)

achieves the goal of a single interaction point with EigenLayer (per AVS).
all calls which push data to EigenLayer core are now forwarded via the ServiceManager(Base)

* fix: use correct addresses in test setup + functions

simple follow-up commit to fix breaking tests from last commit

* chore: restore optimizer_runs to 200

other changes in this PR remove enough from the size
of the RegistryCoordinator contract to make this possible.
200 runs is a wider industry default +
this should slightly decrease gas costs for users.
we can switch back to 100 in the future if we ever need to

* chore: add `setMetadataURI` function to IServiceManager interface

* feat: add operator strategy indexing to service manager (#110)

---------

Co-authored-by: Yash Patil <40046473+ypatil12@users.noreply.github.com>