refactor: singleton cv combining ECDSA and BN254 (#479)
**Motivation:**

Redesigning to have a singleton CertificateVerifier that is deployed per
destination chain where the AVSs utilizing this contract trust the
`globalOperatorTableRoot` is valid and the EigenDA operatorSet which
configures this in a timely basis.

**Modifications:**

- Combined the ECDSACertificateVerifier and the BN254CertificateVerifier
interfaces into CertificateVerifier.
- OperatorSet configuration parameters are now configured on the L1 via
a `CrossChainRegistry` contract and transported by the following
functions.
- `ICertificateVerifier.updateECDSAOperatorTable` or
`ICertificateVerifier.updateBN254OperatorTable`

**Result:**

Standarizing interface with a canonical CertVerifier contract and less
deployment duplication from AVSs.