feat: avs registrar (#484)
**Motivation:**

We want to update the AVS registrar to be minimal & modular. 

**Modifications:**

Add a new `AVSRegistrar` with first class support with the
`KeyRegistrar`. We also have the following modules:
1. `SocketRegistry`
2. `AllowList`

The above modules are _abstract_ and are inherited by `AVSRegistrars` in
the `presets` folder: `AVSRegistrarWithAllowlist` and
`AVSRegistrarWithSocket`.

**Result:**

Simple AVSRegistrar. Note that we are still pending on the
`KeyRegistrar` interface update:
https://github.com/Layr-Labs/eigenlayer-contracts/pull/1421