fix(m-01): add salt to merkle leaf hashing (#527)
**Motivation:**

As part of an audit finding, to protect against [second preimage
attacks](https://flawed.net.nz/2018/02/21/attacking-merkle-trees-with-a-second-preimage-attack/),
we add a salt to the leaf similar to the RewardsCoordinator to
significantly reduce the likelihood of an internal node being used to
produce an unintentional proof.
https://github.com/Layr-Labs/eigenlayer-contracts/pull/1580 in core
already did this on the cert verifier side. We now need to do this for
table calc

**Modifications:**

- Inherit `LeafCalculatorMixin` in `BN254TableCalculatorBase`
- Use `calculateOperatorInfoLeaf`

**Result:**

Secure code